Make hw3
