+++ controllers/auth/index.js
@@ -1,0 +1,6 @@
+const registration = require('./registration');
+const login = require('./login');
+const logout = require('./logout');
+const current = require('./current');
+
+module.exports = { registration, login, logout, current };
