+++ routes/api/index.js
@@ -1,0 +1,4 @@
+const authRouter = require('./auth-router');
+const contactsRouter = require('./contacts-router');
+
+module.exports = { authRouter, contactsRouter };
